refactor(routes): split responses.ts into 3 focused modules (#625)
- responses-passthrough.ts (458 lines): stream/collect translators,
  format adapter, usage extraction, SSE error builders
- responses-compact.ts (199 lines): /compact handler with retry loop
- responses.ts (348 lines): route registration + request assembly

No logic changes — pure structural extraction. All existing tests
pass without modification via re-exports.
